+++ Exchange Storyboard/Services/DateService/DateServiceProtocol.swift
@@ -9,6 +9,5 @@ import Foundation
 
 protocol DateServiceProtocol {
 
-//    func getData(complition: @escaping (Result<[String], Error>) -> Void)
-    func getData(month: Int, complition: @escaping (Result<[String], Error>) -> Void)
+    func getData(month: Double, complition: @escaping (Result<[String], Error>) -> Void)
 }
